fix: use unsigned short for keycode in rshift combo block
The local keycode variable in the RSHIFT combo section (line 276) was
declared as uint8_t but assigned from keycodes[], which is unsigned short.
This caused silent truncation of keycodes > 255. Changed to unsigned short
to match the source type and prevent data loss.

Build: cd keyboard && sudo ./setup_keyboard.sh

+++ keyboard/picocalc_kbd/picocalc_kbd.c
@@ -273,7 +273,7 @@ static void key_report_event(struct kbd_ctx *ctx,
 
   if (ctx->rshift_held && ev->state == KEY_STATE_PRESSED) {
     int i;
-    uint8_t keycode;
+    unsigned short keycode;
     ctx->rshift_used = 1;
 
     if (!ctx->lshift_held) {
